requirements.txt
fastapi
uvicorn
python-multipart
motor
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
google-auth
face-recognition
Pillow
numpy
pydantic
starlette
aiofiles